bizdays' options

bizdays.options defines option parameters used internally in bizdays.

bizdays.options

Format

A list object with methods get and set attached to.

Details

Parameters are stored in bizdays.options using get and set

bizdays.options$set(option.key=value)
bizdays.options$get("option.key")

bizdays supports the following parameter:

default.calendar: the default calendar to be used with the functions: bizdays, bizdayse, adjust.next, adjust.previous, is.bizday, bizseq, offset.

Examples

create.calendar(name = "actual")
bizdays.options$set(default.calendar = "actual")
bizdays("2013-07-12", "2013-07-22")
#> [1] 10