Offsets the given dates by n business days

Returns the given dates offset by the given amount of n business days.

offset(dates, n, cal)

add.bizdays(dates, n, cal)

Arguments

dates

dates to be offset

n

the amount of business days to offset

cal

the calendar's name

Value

Date objects offset by the amount of days defined.

Details

The argument n accepts a sequence of integers and if its length differs from dates' length, the recycle rule is applied to fulfill the gap.

Date types accepted

The argument dates accepts Date objects and any object that returns a valid Date object when passed through as.Date, which include all POSIX* classes and character objects with ISO formatted dates.

Recycle rule

These arguments handle the recycle rule so a vector of dates and a vector of numbers can be provided and once those vectors differs in length the recycle rule is applied.

Examples

offset("2013-01-02", 5, "Brazil/ANBIMA")
#> [1] "2013-01-09"

# Once you have a default calendar set, cal does not need to be provided
bizdays.options$set(default.calendar = "Brazil/ANBIMA")

dates <- seq(as.Date("2013-01-01"), as.Date("2013-01-05"), by = "day")
is.bizday(dates)
#> [1] FALSE  TRUE  TRUE  TRUE FALSE
offset(dates, 1)
#> [1] "2013-01-02" "2013-01-03" "2013-01-04" "2013-01-07" "2013-01-07"